+++ packages/core/src/common/container/auto-bind.ts
@@ -6,16 +6,36 @@ import { AopProxyFactory, ClassFilter } from '../aop/aop-protocol';
 import { ContainerUtil } from './container-util';
 import { Scope } from './scope';
 
-export function autoBind(registry?: interfaces.ContainerModuleCallBack): interfaces.ContainerModule {
+const componentMetadataMap = new Map<string, ComponentMetadata[]>();
+const constantMetadataMap = new Map<string, ConstantOption[]>();
+
+export function autoBindTesting(name: string = 'test'): interfaces.ContainerModule {
+    return autoBind(() => {}, name);
+}
+
+export function autoBind(registry?: interfaces.ContainerModuleCallBack, name?: string): interfaces.ContainerModule {
     return new ContainerModule((bind, unbind, isBound, rebind, ...rest) => {
-        const metadatas: ComponentMetadata[] = Reflect.getMetadata(METADATA_KEY.component, Reflect) || [];
-        for (let index = metadatas.length - 1; index >= 0; index--) {
-            const metadata = metadatas[index];
+        let metadatas: ComponentMetadata[] | undefined;
+        let constantMetadata: ConstantOption[] | undefined;
+        if (name) {
+            metadatas = componentMetadataMap.get(name);
+            constantMetadata = constantMetadataMap.get(name);
+            if (!metadatas) {
+                metadatas = Reflect.getMetadata(METADATA_KEY.component, Reflect) || [];
+                componentMetadataMap.set(name, metadatas!);
+                constantMetadata = Reflect.getMetadata(METADATA_KEY.constantValue, Reflect) || [];
+                constantMetadataMap.set(name, constantMetadata!);
+            }
+        } else {
+            metadatas = Reflect.getMetadata(METADATA_KEY.component, Reflect) || [];
+            constantMetadata = Reflect.getMetadata(METADATA_KEY.constantValue, Reflect) || [];
+        }
+        for (let index = metadatas!.length - 1; index >= 0; index--) {
+            const metadata = metadatas![index];
             resolve(metadata, bind, rebind);
         }
         Reflect.defineMetadata(METADATA_KEY.component, [], Reflect);
-        const constantMetadata: ConstantOption[] = Reflect.getMetadata(METADATA_KEY.constantValue, Reflect) || [];
-        constantMetadata.map(metadata => resolveConstant(metadata, bind, rebind));
+        constantMetadata!.map(metadata => resolveConstant(metadata, bind, rebind));
         Reflect.defineMetadata(METADATA_KEY.constantValue, [], Reflect);
 
         if (registry) {
@@ -43,9 +63,10 @@ function doProxyIfNeed(metadata: ComponentMetadata, target: any) {
 function resolve(metadata: ComponentMetadata, bind: interfaces.Bind, rebind: interfaces.Rebind): void {
     let mid: any;
     const { ids, scope, name, tag, when, proxy, onActivation, target } = metadata;
-    const id = ids.shift()!;
+    const _ids = [...ids];
+    const id = _ids.shift()!;
     if (metadata.rebind) {
-        const id2 = ids.shift();
+        const id2 = _ids.shift();
         mid = rebind(id2 || id).to(target);
     } else {
         mid = bind(id).to(target);
@@ -72,13 +93,13 @@ function resolve(metadata: ComponentMetadata, bind: interfaces.Bind, rebind: int
         mid.onActivation((context: interfaces.Context, t: any) => doProxyIfNeed(metadata, t));
     }
 
-    for (const item of ids) {
+    for (const item of _ids) {
         bind(item).toService(id);
     }
 }
 
 function resolveConstant(metadata: ConstantOption, bind: interfaces.Bind, rebind: interfaces.Rebind): void {
-    const ids = Array.isArray(metadata.id) ? metadata.id : [ metadata.id ];
+    const ids = Array.isArray(metadata.id) ? [...metadata.id] : [ metadata.id ];
     const id = ids.shift();
     if (metadata.rebind) {
         rebind(id!).toConstantValue(metadata.constantValue);